requirements.txt
Flask
Flask-Cors
transformers
openai
requests
python-dotenv
gunicorn
torch
sentence-transformers
Pillow